# backend/target/generated-sources/openapi/src/main/java/com/taxdividend/backend/api/TaxRulesApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.taxdividend.backend.api;

import org.springframework.format.annotation.DateTimeFormat;
import java.time.LocalDate;
import org.springframework.lang.Nullable;
import com.taxdividend.backend.api.dto.TaxRule;
import com.taxdividend.backend.api.dto.TreatyRateResponse;
import java.util.UUID;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-01T10:04:28.893062+01:00[Europe/Zurich]", comments = "Generator version: 7.17.0")
@Validated
public interface TaxRulesApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_FIND_APPLICABLE_RULE = "/tax-rules/applicable";
    /**
     * GET /tax-rules/applicable : Find applicable tax rule
     *
     * @param sourceCountry  (required)
     * @param residenceCountry  (required)
     * @param securityType  (optional, default to EQUITY)
     * @param date  (optional)
     * @return Applicable rule (status code 200)
     *         or No rule found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = TaxRulesApi.PATH_FIND_APPLICABLE_RULE,
        produces = { "application/json" }
    )
    default ResponseEntity<TaxRule> findApplicableRule(
        @NotNull  @Valid @RequestParam(value = "sourceCountry", required = true) String sourceCountry,
        @NotNull  @Valid @RequestParam(value = "residenceCountry", required = true) String residenceCountry,
         @Valid @RequestParam(value = "securityType", required = false, defaultValue = "EQUITY") String securityType,
         @Valid @RequestParam(value = "date", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Nullable LocalDate date
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"residenceCountry\" : \"residenceCountry\", \"effectiveTo\" : \"2000-01-23\", \"sourceCountry\" : \"sourceCountry\", \"standardWithholdingRate\" : 0.8008281904610115, \"notes\" : \"notes\", \"securityType\" : \"securityType\", \"reliefAtSourceAvailable\" : true, \"treatyRate\" : 6.027456183070403, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"refundProcedureAvailable\" : true, \"effectiveFrom\" : \"2000-01-23\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_ALL_TAX_RULES = "/tax-rules";
    /**
     * GET /tax-rules : List tax rules with optional filters
     * Retrieve tax rules with optional filtering by countries, active status, and procedure availability
     *
     * @param sourceCountry Filter by source country (ISO 3166-1 alpha-2 code) (optional)
     * @param residenceCountry Filter by residence country (ISO 3166-1 alpha-2 code) (optional)
     * @param active Filter by active status (rules effective as of today) (optional)
     * @param reliefAtSource Filter by relief at source availability (optional)
     * @param refundProcedure Filter by refund procedure availability (optional)
     * @return List of tax rules (status code 200)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = TaxRulesApi.PATH_GET_ALL_TAX_RULES,
        produces = { "application/json" }
    )
    default ResponseEntity<List<TaxRule>> getAllTaxRules(
        @Pattern(regexp = "^[A-Z]{2}$")  @Valid @RequestParam(value = "sourceCountry", required = false) @Nullable String sourceCountry,
        @Pattern(regexp = "^[A-Z]{2}$")  @Valid @RequestParam(value = "residenceCountry", required = false) @Nullable String residenceCountry,
         @Valid @RequestParam(value = "active", required = false) @Nullable Boolean active,
         @Valid @RequestParam(value = "reliefAtSource", required = false) @Nullable Boolean reliefAtSource,
         @Valid @RequestParam(value = "refundProcedure", required = false) @Nullable Boolean refundProcedure
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"residenceCountry\" : \"residenceCountry\", \"effectiveTo\" : \"2000-01-23\", \"sourceCountry\" : \"sourceCountry\", \"standardWithholdingRate\" : 0.8008281904610115, \"notes\" : \"notes\", \"securityType\" : \"securityType\", \"reliefAtSourceAvailable\" : true, \"treatyRate\" : 6.027456183070403, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"refundProcedureAvailable\" : true, \"effectiveFrom\" : \"2000-01-23\" }, { \"residenceCountry\" : \"residenceCountry\", \"effectiveTo\" : \"2000-01-23\", \"sourceCountry\" : \"sourceCountry\", \"standardWithholdingRate\" : 0.8008281904610115, \"notes\" : \"notes\", \"securityType\" : \"securityType\", \"reliefAtSourceAvailable\" : true, \"treatyRate\" : 6.027456183070403, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"refundProcedureAvailable\" : true, \"effectiveFrom\" : \"2000-01-23\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_TAX_RULE = "/tax-rules/{id}";
    /**
     * GET /tax-rules/{id} : Get tax rule by ID
     *
     * @param id  (required)
     * @return Tax rule details (status code 200)
     *         or Not found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = TaxRulesApi.PATH_GET_TAX_RULE,
        produces = { "application/json" }
    )
    default ResponseEntity<TaxRule> getTaxRule(
        @NotNull  @PathVariable("id") UUID id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"residenceCountry\" : \"residenceCountry\", \"effectiveTo\" : \"2000-01-23\", \"sourceCountry\" : \"sourceCountry\", \"standardWithholdingRate\" : 0.8008281904610115, \"notes\" : \"notes\", \"securityType\" : \"securityType\", \"reliefAtSourceAvailable\" : true, \"treatyRate\" : 6.027456183070403, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"refundProcedureAvailable\" : true, \"effectiveFrom\" : \"2000-01-23\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_TREATY_RATE = "/tax-rules/treaty-rate";
    /**
     * GET /tax-rules/treaty-rate : Get treaty rate
     *
     * @param sourceCountry  (required)
     * @param residenceCountry  (required)
     * @param securityType  (optional, default to EQUITY)
     * @param date  (optional)
     * @return Treaty rate details (status code 200)
     *         or No treaty found (status code 404)
     */
    @RequestMapping(
        method = RequestMethod.GET,
        value = TaxRulesApi.PATH_GET_TREATY_RATE,
        produces = { "application/json" }
    )
    default ResponseEntity<TreatyRateResponse> getTreatyRate(
        @NotNull  @Valid @RequestParam(value = "sourceCountry", required = true) String sourceCountry,
        @NotNull  @Valid @RequestParam(value = "residenceCountry", required = true) String residenceCountry,
         @Valid @RequestParam(value = "securityType", required = false, defaultValue = "EQUITY") String securityType,
         @Valid @RequestParam(value = "date", required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) @Nullable LocalDate date
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"standardRate\" : 0.8008281904610115, \"notes\" : \"notes\", \"reliefAtSourceAvailable\" : true, \"treatyRate\" : 6.027456183070403, \"refundProcedureAvailable\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}