# bff-gateway/target/generated-sources/openapi/src/main/java/com/taxdividend/bff/api/AuthApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.17.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.taxdividend.bff.api;

import org.springframework.lang.Nullable;
import com.taxdividend.bff.model.RegisterRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-01T09:58:25.268465+01:00[Europe/Zurich]", comments = "Generator version: 7.17.0")
@Validated
@Tag(name = "Auth", description = "the Auth API")
public interface AuthApi {

    String PATH_REGISTER = "/auth/register";
    /**
     * POST /auth/register : Register a new user
     *
     * @param registerRequest  (optional)
     * @return Registration initiated (status code 200)
     */
    @Operation(
        operationId = "register",
        summary = "Register a new user",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Registration initiated")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AuthApi.PATH_REGISTER,
        consumes = { "application/json" }
    )
    default Mono<ResponseEntity<Void>> register(
        @Parameter(name = "RegisterRequest", description = "") @Valid @RequestBody(required = false) Mono<RegisterRequest> registerRequest,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(registerRequest).then(Mono.empty());

    }


    String PATH_VERIFY_EMAIL = "/auth/verify";
    /**
     * POST /auth/verify : Verify email
     *
     * @param token  (required)
     * @return Verified (status code 200)
     */
    @Operation(
        operationId = "verifyEmail",
        summary = "Verify email",
        tags = { "Auth" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Verified")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AuthApi.PATH_VERIFY_EMAIL
    )
    default Mono<ResponseEntity<Void>> verifyEmail(
        @NotNull @Parameter(name = "token", description = "", required = true, in = ParameterIn.QUERY) @Valid @RequestParam(value = "token", required = true) String token,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }

}